Moved channel and string extensions in a separate class
