+++ unity/Assets/Tests/AblySandbox/AblyRealtimeExtensions.cs
@@ -9,16 +9,6 @@ namespace Assets.Tests.AblySandbox
 {
     public static class AblyRealtimeExtensions
     {
-        public static string AddRandomSuffix(this string str)
-        {
-            if (str.IsEmpty())
-            {
-                return str;
-            }
-
-            return str + "_" + Guid.NewGuid().ToString("D").Substring(0, 8);
-        }
-
         public static Task<TimeSpan> WaitForState(this AblyRealtime realtime, ConnectionState awaitedState, TimeSpan? waitSpan = null)
         {
             if (realtime.Connection.State == awaitedState)
@@ -33,17 +23,6 @@ namespace Assets.Tests.AblySandbox
             }
 
             return connectionAwaiter.Wait();
-        }
-
-        public static async Task WaitForState(this IRealtimeChannel channel, ChannelState awaitedState = ChannelState.Attached, TimeSpan? waitSpan = null)
-        {
-            var channelAwaiter = new ChannelAwaiter(channel, awaitedState);
-            var timespan = waitSpan.GetValueOrDefault(TimeSpan.FromSeconds(5));
-            Result<bool> result = await channelAwaiter.WaitAsync(timespan);
-            if (result.IsFailure)
-            {
-                throw new Exception($"Channel '{channel.Name}' did not reach '{awaitedState}' in {timespan.TotalSeconds} seconds. Current state {channel.State}");
-            }
         }
 
         public static Task WaitForState(this IRealtimeClient realtime, ConnectionState awaitedState = ConnectionState.Connected, TimeSpan? waitSpan = null)
@@ -88,37 +67,6 @@ namespace Assets.Tests.AblySandbox
 #pragma warning restore CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
 
             await taskAwaiter.Task;
-        }
-
-
-        public static async Task<bool> WaitSync(this Presence presence, TimeSpan? waitSpan = null)
-        {
-            TaskCompletionSource<bool> taskCompletionSource = new TaskCompletionSource<bool>();
-            var inProgress = presence.IsSyncInProgress;
-            if (inProgress == false)
-            {
-                return true;
-            }
-
-            void OnPresenceOnSyncCompleted(object sender, EventArgs e)
-            {
-                presence.SyncCompleted -= OnPresenceOnSyncCompleted;
-                taskCompletionSource.SetResult(true);
-            }
-
-            presence.SyncCompleted += OnPresenceOnSyncCompleted;
-            var timeout = waitSpan ?? TimeSpan.FromSeconds(2);
-            var waitTask = Task.Delay(timeout);
-
-            // Do some waiting. Either the sync will complete or the timeout will finish.
-            // if it is the timeout first then we throw an exception. This way we do go down a rabbit hole
-            var firstTask = await Task.WhenAny(waitTask, taskCompletionSource.Task);
-            if (waitTask == firstTask)
-            {
-                throw new Exception("WaitSync timed out after: " + timeout.TotalSeconds + " seconds");
-            }
-
-            return true;
         }
     }
 }
